Fix a crucial bug (wrong type used for size) in Serialization<std::set<...>>::save()

+++ include/diy/serialization.hpp
@@ -235,7 +235,7 @@ namespace diy
 
     static void         save(BinaryBuffer& bb, const Set& m)
     {
-      unsigned s = m.size();
+      size_t s = m.size();
       diy::save(bb, s);
       for (typename std::set<T>::const_iterator it = m.begin(); it != m.end(); ++it)
         diy::save(bb, *it);
